hello_world/requirements.txt
GoogleAppEngineCloudStorageClient==1.9.22.1
